simplified benchmarks compilation
