+++ tests/benchmark/spaces_benchmarks/bm_spaces.h
@@ -8,7 +8,6 @@
 #include <random>
 #include <unistd.h>
 #include "VecSim/utils/arr_cpp.h"
-#include "VecSim/spaces/space_includes.h"
 #include "VecSim/spaces/space_aux.h"
 
 #include "bm_spaces_class.h"
