+++ tests/benchmark/spaces_benchmarks/bm_spaces_class.h
@@ -8,7 +8,6 @@
 #include <random>
 #include <unistd.h>
 #include "VecSim/utils/arr_cpp.h"
-#include "VecSim/spaces/space_includes.h"
 #include "VecSim/spaces/space_aux.h"
 
 class BM_VecSimSpaces : public benchmark::Fixture {
@@ -25,3 +24,24 @@ public:
     void SetUp(const ::benchmark::State &state);
     void TearDown(const ::benchmark::State &state);
 };
+
+BM_VecSimSpaces::BM_VecSimSpaces() {
+    rng.seed(47);
+    opt = getArchitectureOptimization();
+}
+
+void BM_VecSimSpaces::SetUp(const ::benchmark::State &state) {
+    dim = state.range(0);
+    v1 = new DATA_TYPE[dim];
+    v2 = new DATA_TYPE[dim];
+    std::uniform_real_distribution<double> distrib(-1.0, 1.0);
+    for (size_t i = 0; i < dim; i++) {
+        v1[i] = (DATA_TYPE)distrib(rng);
+        v2[i] = (DATA_TYPE)distrib(rng);
+    }
+}
+
+void BM_VecSimSpaces::TearDown(const ::benchmark::State &state) {
+    delete v1;
+    delete v2;
+}
