--- tests/benchmark/spaces_benchmarks/bm_spaces_class_fp64.cpp
@@ -1,9 +1,0 @@
-/*
- *Copyright Redis Ltd. 2021 - present
- *Licensed under your choice of the Redis Source Available License 2.0 (RSALv2) or
- *the Server Side Public License v1 (SSPLv1).
- */
-
-#define DATA_TYPE double
-
-#include "bm_spaces_class_definitions.h"
